requirements.txt
streamlit
opencv-python
numpy
pillow
gspread
oauth2client
